backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
redis==5.0.4
python-multipart==0.0.9
numpy==1.26.4
python-dotenv==1.0.1
celery==5.4.0
aioredis==2.0.1
greenlet==3.0.3
